refactor: centralize iconify icon definitions
