--- src/component/layout/headerIcons.ts
@@ -1,19 +1,0 @@
-import type { IconifyIcon } from '@iconify/types'
-
-export const iconSearch: IconifyIcon = {
-  width: 24,
-  height: 24,
-  body: '<path fill="currentColor" d="m15.5 14h-.79l-.28-.27a6.5 6.5 0 1 0-.71.71l.27.28v.79L20 21.5L21.5 20zm-6 0A4.5 4.5 0 1 1 14 9.5A4.505 4.505 0 0 1 9.5 14"/>',
-}
-
-export const iconSun: IconifyIcon = {
-  width: 24,
-  height: 24,
-  body: '<path fill="currentColor" d="M6.76 4.84L5.35 3.43L3.93 4.84l1.42 1.42zm10.48 0l1.41 1.42l1.42-1.42l-1.42-1.41zM12 4a1 1 0 0 0 1-1V1h-2v2a1 1 0 0 0 1 1m0 16a1 1 0 0 0-1 1v2h2v-2a1 1 0 0 0-1-1M4 11H2v2h2a1 1 0 0 0 0-2m18 0h-2a1 1 0 0 0 0 2h2zM6.76 19.16l-1.41 1.41l1.41 1.42l1.42-1.42zm10.48 0l1.42 1.41l1.41-1.41l-1.41-1.42zM12 7a5 5 0 1 0 0 10a5 5 0 0 0 0-10"/>',
-}
-
-export const iconMoon: IconifyIcon = {
-  width: 24,
-  height: 24,
-  body: '<path fill="currentColor" d="M9.37 5.51A7 7 0 0 0 18.49 14.63A8 8 0 1 1 9.37 5.51"/>',
-}
